refactor: split extension root-route + polish embed iframe loading
Breaks the 898-line root-route.tsx into focused modules under
routes/root/ (chrome-tab, detect-target, field-values, lookups, types,
views) and removes the now-unused parsed-card and record-detail-skeleton
components. Embed shell switches to a single h-screen scroll container,
the record view defers render until useRecord resolves so the iframe
shows one continuous loader, and the header gains an inline "Open"
affordance.
