+++ apps/extension/src/iframe/routes/root-route.tsx
@@ -1,23 +1,21 @@
 // apps/extension/src/iframe/routes/root-route.tsx
 
-import { Button } from '@auxx/ui/components/button'
-import { ChevronRight } from 'lucide-react'
 import { useCallback, useEffect, useRef, useState } from 'react'
-import type { PageOperation } from '../../lib/messaging'
-import type { ParsedCompany, ParsedPerson, ParseResult } from '../../lib/parsers/types'
-import { ParsedCard } from '../components/parsed-card'
+import type { ParseResult } from '../../lib/parsers/types'
 import { useRouteStack } from '../hooks/use-route-stack'
+import { createRecord, reportParserHealth, type SessionResponse, TrpcCallError } from '../trpc'
+import { invokeOnPage, OWN_TAB_ID, readOwnTab } from './root/chrome-tab'
+import { detectTarget, facebookContactInfoUrl, readGenericPage } from './root/detect-target'
 import {
-  BASE_URL,
-  createRecord,
-  type LookupByFieldCandidate,
-  lookupByField,
-  reportParserHealth,
-  type SessionResponse,
-  TrpcCallError,
-  uploadAvatarFromUrl,
-} from '../trpc'
-import { instanceIdFromRecordId } from './types'
+  buildAvatarFieldValue,
+  buildCompanyFieldValues,
+  buildContactFieldValues,
+  companyToPerson,
+  personToCompany,
+} from './root/field-values'
+import { findExistingByGenericPage, findExistingInEntity } from './root/lookups'
+import type { EntityType, ExistingMatch, Phase, Target } from './root/types'
+import { PhaseView } from './root/views'
 
 /**
  * The capture-from-this-page flow. Lives inside the iframe shell's main
@@ -29,515 +27,12 @@ import { instanceIdFromRecordId } from './types'
  *   - rendering parsed fields + Save / Open CTAs
  *
  * Signed-out / no-org cases short-circuit with the existing CTAs.
- */
-
-type EntityType = 'contact' | 'company'
-
-type Target = {
-  host: SupportedHost
-  entityType: EntityType
-  parseOp: PageOperation
-}
-
-type Phase =
-  | { kind: 'loading' }
-  | { kind: 'signed-out' }
-  | { kind: 'no-org' }
-  | { kind: 'generic-site'; page: GenericPage | null; existingRecordId: string | null }
-  | { kind: 'saving-generic'; page: GenericPage }
-  | { kind: 'parsing'; target: Target }
-  | {
-      kind: 'ready'
-      target: Target
-      person: ParsedPerson | null
-      company: ParsedCompany | null
-      /**
-       * All existing records matching the parse identity, across both
-       * contact and company. Folk's "N similar found" list — clicking a
-       * row pushes the corresponding detail route. Sorted with same-
-       * entity-type-as-target hits first so the most likely hit shows up
-       * top.
-       */
-      existingMatches: ExistingMatch[]
-      /**
-       * Lookup status. `loading` after parse but before the
-       * `record.lookupByField` round-trip resolves. We need this
-       * distinct from "loaded with zero matches" so the view doesn't
-       * briefly render the capture form (no matches → ReadyToSaveView)
-       * before the matches arrive — which is what produced the
-       * "flash of Add-anyway view on back-navigation" bug.
-       */
-      matchesStatus: 'loading' | 'loaded'
-      /**
-       * Which Save button is mid-flight, if any. Drives the per-button
-       * `loading` state without flipping the whole panel into the
-       * "parsing" phase (which would unmount the parsed card).
-       */
-      savingAs: EntityType | null
-    }
-  | { kind: 'needs-fb-contact-info'; profileUrl: string }
-  | { kind: 'error'; message: string }
-
-/**
- * The active tab's URL + title, normalized. Captured once on boot for the
- * generic-site capture view. `hostname` is lowercased and www-stripped so
- * it can double as a company_domain value on save.
- */
-type GenericPage = {
-  url: string
-  title: string
-  hostname: string
-}
-
-function readGenericPage(tab: chrome.tabs.Tab | null): GenericPage | null {
-  const rawUrl = tab?.url
-  if (!rawUrl) return null
-  try {
-    const u = new URL(rawUrl)
-    if (u.protocol !== 'http:' && u.protocol !== 'https:') return null
-    const hostname = u.hostname.replace(/^www\./, '').toLowerCase()
-    if (!hostname) return null
-    const title = (tab?.title ?? '').trim() || hostname
-    return { url: rawUrl, title, hostname }
-  } catch {
-    return null
-  }
-}
-
-type SupportedHost = 'gmail' | 'linkedin' | 'sales-navigator' | 'twitter' | 'facebook' | 'instagram'
-
-const HOST_LABEL: Record<SupportedHost, string> = {
-  gmail: 'Gmail',
-  linkedin: 'LinkedIn',
-  'sales-navigator': 'Sales Navigator',
-  twitter: 'X',
-  facebook: 'Facebook',
-  instagram: 'Instagram',
-}
-
-/**
- * Top-level x.com / twitter.com paths that aren't usernames. Without this
- * guard, the iframe would try to parse a profile out of `x.com/home` and
- * silently fail.
- */
-const TWITTER_RESERVED_PATHS = new Set([
-  'home',
-  'explore',
-  'notifications',
-  'messages',
-  'i',
-  'compose',
-  'settings',
-  'tos',
-  'privacy',
-  'about',
-  'search',
-  'login',
-  'signup',
-  'logout',
-  'account',
-  'jobs',
-  'premium_sign_up',
-])
-
-/**
- * Top-level facebook.com paths that aren't profiles. `profile.php` is
- * whitelisted by falling through — we key dedup off the `id` query param.
- */
-const FACEBOOK_RESERVED_PATHS = new Set([
-  'home',
-  'marketplace',
-  'groups',
-  'watch',
-  'events',
-  'messages',
-  'friends',
-  'bookmarks',
-  'notifications',
-  'settings',
-  'gaming',
-  'reel',
-  'pages',
-  'stories',
-  'policies',
-  'help',
-  'business',
-  'login',
-  'signup',
-  'logout',
-])
-
-/**
- * Top-level instagram.com paths that aren't profiles. `/<username>/followers`,
- * `/<username>/tagged`, etc. aren't listed here — the first segment is the
- * username, so the parser's `h2 === url slug` guard gates parse success.
- */
-const INSTAGRAM_RESERVED_PATHS = new Set([
-  'accounts',
-  'direct',
-  'explore',
-  'p',
-  'reel',
-  'reels',
-  'stories',
-  'tv',
-  'tags',
-  'locations',
-  'legal',
-  'about',
-  'developer',
-  'press',
-  'web',
-  'emails',
-  'privacy',
-  'session',
-  'challenge',
-  'oauth',
-  'api',
-])
-
-/**
- * Returns the deep-link to the Contact and Basic Info sub-tab of the current
- * Facebook profile, or null when the URL already points there. The parser
- * only succeeds inside that tab so we route the user there via a hint before
- * dispatching.
- */
-function facebookContactInfoUrl(u: URL): string | null {
-  if (u.pathname.includes('about_contact_and_basic_info')) return null
-  if (u.searchParams.get('sk') === 'about_contact_and_basic_info') return null
-  if (u.pathname.includes('profile.php')) {
-    const id = u.searchParams.get('id')
-    if (!id) return null
-    return `https://www.facebook.com/profile.php?id=${id}&sk=about_contact_and_basic_info`
-  }
-  const vanity = u.pathname.split('/').filter(Boolean)[0]
-  if (!vanity) return null
-  return `https://www.facebook.com/${vanity}/about_contact_and_basic_info`
-}
-
-function detectTarget(url: string | undefined): Target | null {
-  if (!url) return null
-  try {
-    const u = new URL(url)
-    if (u.hostname === 'mail.google.com') {
-      return { host: 'gmail', entityType: 'contact', parseOp: 'parseGmail' }
-    }
-    if (u.hostname === 'www.linkedin.com') {
-      if (u.pathname.startsWith('/sales/')) {
-        return {
-          host: 'sales-navigator',
-          entityType: 'contact',
-          parseOp: 'parseSalesNavigator',
-        }
-      }
-      if (u.pathname.startsWith('/company/')) {
-        return {
-          host: 'linkedin',
-          entityType: 'company',
-          parseOp: 'parseLinkedInCompany',
-        }
-      }
-      return { host: 'linkedin', entityType: 'contact', parseOp: 'parseLinkedIn' }
-    }
-    if (u.hostname === 'twitter.com' || u.hostname === 'x.com' || u.hostname === 'www.x.com') {
-      if (u.pathname.startsWith('/search') || u.pathname.startsWith('/i/lists/')) {
-        return { host: 'twitter', entityType: 'contact', parseOp: 'parseTwitterSearch' }
-      }
-      const firstSegment = u.pathname.split('/')[1] ?? ''
-      if (!firstSegment || TWITTER_RESERVED_PATHS.has(firstSegment)) return null
-      return { host: 'twitter', entityType: 'contact', parseOp: 'parseTwitterProfile' }
-    }
-    if (u.hostname === 'www.facebook.com' || u.hostname === 'facebook.com') {
-      const firstSegment = u.pathname.split('/')[1] ?? ''
-      if (firstSegment !== 'profile.php' && FACEBOOK_RESERVED_PATHS.has(firstSegment)) return null
-      if (!firstSegment) return null
-      // entityType is provisional — the parser self-detects and the
-      // parse-result handler re-routes to 'company' when the result
-      // populates the companies array. See the run() effect below.
-      return { host: 'facebook', entityType: 'contact', parseOp: 'parseFacebook' }
-    }
-    if (u.hostname === 'www.instagram.com' || u.hostname === 'instagram.com') {
-      const firstSegment = u.pathname.split('/')[1] ?? ''
-      if (!firstSegment) return null
-      if (INSTAGRAM_RESERVED_PATHS.has(firstSegment)) return null
-      return { host: 'instagram', entityType: 'contact', parseOp: 'parseInstagramProfile' }
-    }
-    return null
-  } catch {
-    return null
-  }
-}
-
-/**
- * The iframe's own tab id, written into `iframe.src` as `?tabId=N` by the
- * MAIN-world inject script (see `background.ts`). Captured at module load
- * because window.location is fixed for the iframe's lifetime.
  *
- * Reading the iframe's *own* tab — instead of `chrome.tabs.query({active})`
- * — is what keeps a background-tab iframe from re-rendering with the
- * focused tab's URL when something else triggers a re-boot.
+ * Supporting code (types, host detection, Chrome plumbing, field-value
+ * builders, lookups, view components) lives under `./root/`. This file
+ * keeps just the orchestrator: the `Phase` state machine + effects + the
+ * Save handlers.
  */
-const OWN_TAB_ID: number | null = (() => {
-  const raw = new URLSearchParams(window.location.search).get('tabId')
-  if (!raw) return null
-  const n = Number(raw)
-  return Number.isFinite(n) ? n : null
-})()
-
-async function readOwnTab(): Promise<chrome.tabs.Tab | null> {
-  if (OWN_TAB_ID === null) {
-    // Fallback for older injected frames (no tabId param) — preserves the
-    // prior behaviour. Should be unreachable in practice once users reload.
-    try {
-      const [tab] = await chrome.tabs.query({ active: true, lastFocusedWindow: true })
-      return tab ?? null
-    } catch {
-      return null
-    }
-  }
-  try {
-    return await chrome.tabs.get(OWN_TAB_ID)
-  } catch {
-    return null
-  }
-}
-
-async function invokeOnPage<T = unknown>(operation: PageOperation): Promise<T | null> {
-  return new Promise((resolve) => {
-    chrome.runtime.sendMessage({ type: 'invoke', operation, args: [] }, (response) => {
-      if (chrome.runtime.lastError) {
-        resolve(null)
-        return
-      }
-      if (!response?.ok) {
-        resolve(null)
-        return
-      }
-      resolve((response.value as T) ?? null)
-    })
-  })
-}
-
-// ─── Field-value mappers ───────────────────────────────────────
-
-function buildContactFieldValues(person: ParsedPerson): Record<string, unknown> {
-  return {
-    // full_name is FieldType.NAME — a server-derived view over first_name +
-    // last_name (see field-value-helpers.ts resolveNameFieldDisplayValue /
-    // field-value-queries.ts NAME hydration). Writing to it directly would
-    // fail validation (NAME expects { firstName, lastName }, not a string),
-    // and it's redundant — setting first/last populates the displayName
-    // column and synthesises the NAME value on read.
-    first_name: person.firstName,
-    last_name: person.lastName,
-    primary_email: person.primaryEmail,
-    phone: person.phone,
-    notes: person.notes,
-    // externalId is multi-value (options.multi=true in the registry). Send
-    // the capture's identifier as a single-element array so the server
-    // writes one FieldValue row; future recaptures append via mode: 'add'.
-    external_id: [person.externalId],
-    // contact_avatar is a FILE field — populated separately via
-    // buildAvatarFieldValue() which resolves the URL to an asset ref.
-  }
-}
-
-function buildCompanyFieldValues(company: ParsedCompany): Record<string, unknown> {
-  return {
-    company_name: company.name,
-    company_domain: company.domain,
-    company_notes: company.notes,
-    company_x_follower_count: company.xFollowerCount,
-    // externalId is multi-value — see buildContactFieldValues.
-    external_id: [company.externalId],
-    // company_logo is a FILE field — populated separately via
-    // buildAvatarFieldValue() which resolves the URL to an asset ref.
-  }
-}
-
-/**
- * Convert a parsed person into company-shaped data. Used when the user
- * clicks the secondary "Save as company" button on a profile that Auxx
- * detected as a contact (e.g. a Twitter org account or an ambiguous page).
- * The externalId is passed through as-is — the same URL always produces the
- * same identifier, so re-captures will still dedup against the original.
- */
-function personToCompany(person: ParsedPerson | null): ParsedCompany | null {
-  if (!person) return null
-  const joinedFromParts = [person.firstName, person.lastName].filter(Boolean).join(' ')
-  const name = person.fullName ?? (joinedFromParts.length > 0 ? joinedFromParts : undefined)
-  return {
-    name,
-    avatarUrl: person.avatarUrl,
-    notes: person.notes,
-    externalId: person.externalId,
-    xFollowerCount: person.xFollowerCount,
-  }
-}
-
-/** Inverse of `personToCompany` — for the "Save as contact" button on a company parse. */
-function companyToPerson(company: ParsedCompany | null): ParsedPerson | null {
-  if (!company) return null
-  const parts = (company.name ?? '').split(/\s+/).filter(Boolean)
-  const firstName = parts[0]
-  const lastName = parts.length > 1 ? parts.slice(1).join(' ') : undefined
-  return {
-    firstName,
-    lastName,
-    fullName: company.name,
-    avatarUrl: company.avatarUrl,
-    notes: company.notes,
-    externalId: company.externalId,
-  }
-}
-
-// FILE fields are keyed by systemAttribute in the resource registry.
-const AVATAR_SYSTEM_ATTR: Record<EntityType, string> = {
-  contact: 'contact_avatar',
-  company: 'company_logo',
-}
-
-async function buildAvatarFieldValue(
-  entityType: EntityType,
-  avatarUrl: string | undefined
-): Promise<Record<string, unknown>> {
-  if (!avatarUrl) return {}
-  try {
-    const { ref } = await uploadAvatarFromUrl({ url: avatarUrl, entityType })
-    return { [AVATAR_SYSTEM_ATTR[entityType]]: { ref } }
-  } catch (err) {
-    // Avatar upload is best-effort — fall through to saving the record
-    // without the image. LinkedIn CDN URLs expire; a 403 here is expected.
-    console.warn('[auxx] avatar upload failed, saving without', err)
-    return {}
-  }
-}
-
-// ─── Dedup lookups ─────────────────────────────────────────────
-//
-// Single round-trip priority lookup via `record.lookupByField`:
-// externalId first (exact match on the capture's source+id), then
-// primary_email / company_domain as a wider-net fallback. See
-// plans/folk/21-record-lookup-by-field.md §"Wire shape" for why we send
-// the whole list in one call rather than two serial searches.
-//
-// Limit raised to 5 so the iframe can show folk's "N similar found" list
-// instead of pretending the first hit is the only one.
-
-const LOOKUP_LIMIT = 5
-
-/**
- * One existing-match row for the "N similar found" list. Carries the
- * composite recordId (what `record.getById` takes) plus the entity type so
- * the click handler can push the right route variant without re-deriving
- * either from the recordId's entity-definition prefix.
- */
-type ExistingMatch = {
-  recordId: string
-  entityType: EntityType
-  matchedBy: { systemAttribute: string; value: string }
-}
-
-async function findExistingByPerson(person: ParsedPerson): Promise<ExistingMatch[]> {
-  const candidates: LookupByFieldCandidate[] = [
-    { systemAttribute: 'external_id', value: person.externalId },
-  ]
-  if (person.primaryEmail) {
-    candidates.push({ systemAttribute: 'primary_email', value: person.primaryEmail })
-  }
-  try {
-    const result = await lookupByField({
-      entityDefinitionId: 'contact',
-      candidates,
-      limit: LOOKUP_LIMIT,
-    })
-    return result.items.map((item) => ({
-      recordId: item.recordId,
-      entityType: 'contact',
-      matchedBy: {
-        systemAttribute: item.matchedBy.systemAttribute,
-        value: String(item.matchedBy.value ?? ''),
-      },
-    }))
-  } catch {
-    return []
-  }
-}
-
-async function findExistingByCompany(company: ParsedCompany): Promise<ExistingMatch[]> {
-  const candidates: LookupByFieldCandidate[] = [
-    { systemAttribute: 'external_id', value: company.externalId },
-  ]
-  if (company.domain) {
-    candidates.push({ systemAttribute: 'company_domain', value: company.domain })
-  }
-  try {
-    const result = await lookupByField({
-      entityDefinitionId: 'company',
-      candidates,
-      limit: LOOKUP_LIMIT,
-    })
-    return result.items.map((item) => ({
-      recordId: item.recordId,
-      entityType: 'company',
-      matchedBy: {
-        systemAttribute: item.matchedBy.systemAttribute,
-        value: String(item.matchedBy.value ?? ''),
-      },
-    }))
-  } catch {
-    return []
-  }
-}
-
-/**
- * Look up existing records in a specific entity type, synthesizing the
- * cross-entity candidate via personToCompany / companyToPerson when the
- * parse result only carries the other shape. Twitter profiles, for
- * example, always parse as a person — but the user may have hit the
- * secondary "Save as company" button on a prior visit, writing a company
- * row with the same externalId. Without this cross-lookup, the iframe
- * would offer Save again on re-visit.
- */
-async function findExistingInEntity(
-  entityType: EntityType,
-  person: ParsedPerson | null,
-  company: ParsedCompany | null
-): Promise<ExistingMatch[]> {
-  if (entityType === 'contact') {
-    const p = person ?? companyToPerson(company)
-    return p ? findExistingByPerson(p) : []
-  }
-  const c = company ?? personToCompany(person)
-  return c ? findExistingByCompany(c) : []
-}
-
-async function findExistingByGenericPage(page: GenericPage): Promise<string | null> {
-  // Generic-site capture stays single-hit — the URL→company mapping is
-  // 1:1 (we key on hostname), so a list view would only ever have one
-  // row. Single string return preserves the existing call sites.
-  const candidates: LookupByFieldCandidate[] = [
-    { systemAttribute: 'external_id', value: `website:${page.hostname}` },
-    { systemAttribute: 'company_domain', value: page.hostname },
-  ]
-  try {
-    const result = await lookupByField({ entityDefinitionId: 'company', candidates })
-    const hit = result.items[0]?.recordId
-    return hit ? instanceIdFromRecordId(hit) : null
-  } catch {
-    return null
-  }
-}
-
-function recordDeepLink(entityType: EntityType, recordId: string): string {
-  return entityType === 'contact'
-    ? `${BASE_URL}/app/contacts/${recordId}`
-    : `${BASE_URL}/app/companies/${recordId}`
-}
-
-function loginUrl(): string {
-  return `${BASE_URL}/login`
-}
 
 type Props = {
   session: SessionResponse
@@ -780,13 +275,7 @@ export function RootRoute({ session }: Props) {
       })
       navigateToDetail('company', created.recordId)
     } catch (err) {
-      const message =
-        err instanceof TrpcCallError
-          ? `Save failed (${err.code ?? err.httpStatus ?? 'error'}): ${err.message}`
-          : err instanceof Error
-            ? err.message
-            : 'Save failed.'
-      setPhase({ kind: 'error', message })
+      setPhase({ kind: 'error', message: saveErrorMessage(err) })
     }
   }, [phase, navigateToDetail])
 
@@ -796,7 +285,7 @@ export function RootRoute({ session }: Props) {
       if (phase.savingAs) return
       // Snapshot the ready phase so the avatar upload + create can run
       // without us flipping back to "parsing" (which would unmount the
-      // ParsedCard and lose the in-flight `savingAs` flag).
+      // suggestions list and lose the in-flight `savingAs` flag).
       const ready = phase
       setPhase({ ...ready, savingAs: entityType })
       try {
@@ -826,13 +315,7 @@ export function RootRoute({ session }: Props) {
         })
         navigateToDetail('company', created.recordId)
       } catch (err) {
-        const message =
-          err instanceof TrpcCallError
-            ? `Save failed (${err.code ?? err.httpStatus ?? 'error'}): ${err.message}`
-            : err instanceof Error
-              ? err.message
-              : 'Save failed.'
-        setPhase({ kind: 'error', message })
+        setPhase({ kind: 'error', message: saveErrorMessage(err) })
       }
     },
     [phase, navigateToDetail]
@@ -847,341 +330,10 @@ export function RootRoute({ session }: Props) {
   )
 }
 
-function PhaseView({
-  phase,
-  onSaveAs,
-  onSaveGenericCompany,
-}: {
-  phase: Phase
-  onSaveAs: (entityType: EntityType) => void
-  onSaveGenericCompany: () => void
-}) {
-  switch (phase.kind) {
-    case 'loading':
-      return <p className='text-sm text-muted-foreground'>Loading…</p>
-
-    case 'signed-out':
-      return (
-        <Empty
-          title='Sign in to Auxx'
-          body='Connect to your Auxx workspace to save contacts directly from this page.'
-          action={{ label: 'Sign in', href: loginUrl() }}
-        />
-      )
-
-    case 'no-org':
-      return (
-        <Empty
-          title='No active organization'
-          body='Pick a workspace inside Auxx to start saving contacts here.'
-          action={{ label: 'Open Auxx', href: BASE_URL }}
-        />
-      )
-
-    case 'generic-site':
-      return (
-        <GenericSiteView
-          page={phase.page}
-          existingRecordId={phase.existingRecordId}
-          onSaveCompany={onSaveGenericCompany}
-        />
-      )
-
-    case 'saving-generic':
-      return <p className='text-sm text-muted-foreground'>Saving {phase.page.hostname}…</p>
-
-    case 'parsing':
-      return (
-        <p className='text-sm text-muted-foreground'>Reading {HOST_LABEL[phase.target.host]}…</p>
-      )
-
-    case 'ready': {
-      const target = phase.target
-      const subject = target.entityType === 'contact' ? phase.person : phase.company
-      if (!subject) {
-        return (
-          <Empty
-            title={`No ${target.entityType} found on this ${HOST_LABEL[target.host]} page`}
-            body={
-              target.entityType === 'contact'
-                ? 'Try opening a thread or profile with a recipient or person details.'
-                : 'Try opening a company profile.'
-            }
-          />
-        )
-      }
-      // Defer rendering the matches/capture decision until the lookup
-      // round-trip resolves — without this, the empty-matches branch
-      // renders briefly before matches arrive (visible as a flash of
-      // the capture view on back-navigation from a detail route).
-      if (phase.matchesStatus === 'loading') {
-        return <p className='text-sm text-muted-foreground'>Checking for existing records…</p>
-      }
-      // Two-step UX when there are existing matches: show the "N similar
-      // found" list first (no parsed card, no save buttons) so the user
-      // doesn't accidentally double-save. They can opt into the capture
-      // view via "Add anyway?" — that path renders the parsed card +
-      // save buttons, identical to the no-matches path. Folk's flow.
-      return <ReadyView phase={phase} onSaveAs={onSaveAs} />
-    }
-
-    case 'needs-fb-contact-info':
-      return (
-        <Empty
-          title='Open Contact and Basic Info'
-          body='Facebook only exposes profile details on the About → Contact and Basic Info tab.'
-          action={{ label: 'Open tab', href: phase.profileUrl }}
-        />
-      )
-
-    case 'error':
-      return (
-        <Empty
-          title='Something went wrong'
-          body={phase.message}
-          action={{ label: 'Open Auxx', href: BASE_URL }}
-        />
-      )
+function saveErrorMessage(err: unknown): string {
+  if (err instanceof TrpcCallError) {
+    return `Save failed (${err.code ?? err.httpStatus ?? 'error'}): ${err.message}`
   }
-}
-
-function GenericSiteView({
-  page,
-  existingRecordId,
-  onSaveCompany,
-}: {
-  page: GenericPage | null
-  existingRecordId: string | null
-  onSaveCompany: () => void
-}) {
-  if (!page) {
-    return (
-      <Empty
-        title='No page detected'
-        body='Open Auxx to manage your records, or navigate to a page you want to capture.'
-        action={{ label: 'Open Auxx', href: BASE_URL }}
-      />
-    )
-  }
-  if (existingRecordId) {
-    return (
-      <div className='space-y-4'>
-        <h2 className='text-sm text-muted-foreground'>Already in Auxx</h2>
-        <div className='space-y-1 rounded-md border border-border bg-muted/30 px-3 py-2'>
-          <p className='truncate text-sm font-medium' title={page.title}>
-            {page.title}
-          </p>
-          <p className='truncate text-xs text-muted-foreground' title={page.url}>
-            {page.hostname}
-          </p>
-        </div>
-        <Button asChild className='w-full'>
-          <a href={recordDeepLink('company', existingRecordId)} target='_blank' rel='noreferrer'>
-            Open in Auxx
-          </a>
-        </Button>
-      </div>
-    )
-  }
-  return (
-    <div className='space-y-4'>
-      <div className='space-y-1'>
-        <h2 className='text-base font-medium'>Save this page</h2>
-        <p className='text-sm text-muted-foreground'>
-          This page isn't a profile Auxx can parse yet. You can still capture it as a company.
-        </p>
-      </div>
-      <div className='space-y-1 rounded-md border border-border bg-muted/30 px-3 py-2'>
-        <p className='truncate text-sm font-medium' title={page.title}>
-          {page.title}
-        </p>
-        <p className='truncate text-xs text-muted-foreground' title={page.url}>
-          {page.hostname}
-        </p>
-      </div>
-      <Button onClick={onSaveCompany} className='w-full'>
-        Save as company
-      </Button>
-      <Button asChild variant='outline' className='w-full'>
-        <a href={BASE_URL} target='_blank' rel='noreferrer'>
-          Open Auxx
-        </a>
-      </Button>
-    </div>
-  )
-}
-
-/**
- * Top-level switch for the `ready` phase. Reads the root sub-view off the
- * route stack — `matches` shows the existing-matches list (default when
- * any matches were found), `capture` shows the parsed card + save buttons
- * (folk's "Add anyway?" path).
- *
- * Putting this in the route stack (instead of local useState) lets the
- * header's back chevron handle "back to matches" for free, and pops on
- * parse-identity change so a SPA navigation to a different profile resets
- * to the matches list for that profile.
- */
-function ReadyView({
-  phase,
-  onSaveAs,
-}: {
-  phase: Extract<Phase, { kind: 'ready' }>
-  onSaveAs: (entityType: EntityType) => void
-}) {
-  const { top, push, pop } = useRouteStack()
-  const parseKey = phase.person?.externalId ?? phase.company?.externalId ?? ''
-  // biome-ignore lint/correctness/useExhaustiveDependencies: parseKey is the trigger; pop/top are read live.
-  useEffect(() => {
-    if (top.kind === 'root' && top.view === 'capture') pop()
-  }, [parseKey])
-
-  const isCaptureView = top.kind === 'root' && top.view === 'capture'
-
-  if (phase.existingMatches.length > 0 && !isCaptureView) {
-    return (
-      <MatchesOnlyView
-        matches={phase.existingMatches}
-        onAddAnyway={() => push({ kind: 'root', view: 'capture' })}
-      />
-    )
-  }
-  return <ReadyToSaveView phase={phase} onSaveAs={onSaveAs} />
-}
-
-function ReadyToSaveView({
-  phase,
-  onSaveAs,
-}: {
-  phase: Extract<Phase, { kind: 'ready' }>
-  onSaveAs: (entityType: EntityType) => void
-}) {
-  // Primary = the parser's best guess. Secondary = the other — some profiles
-  // (brand accounts, solo-founder pages, verified orgs) are genuinely
-  // ambiguous, so we always let the user pick.
-  const primary = phase.target.entityType
-  const secondary: EntityType = primary === 'contact' ? 'company' : 'contact'
-  const isSaving = phase.savingAs !== null
-  return (
-    <div className='space-y-4'>
-      {phase.person && <ParsedCard person={phase.person} />}
-      {phase.company && <ParsedCard company={phase.company} />}
-      <Button
-        onClick={() => onSaveAs(primary)}
-        className='w-full'
-        loading={phase.savingAs === primary}
-        loadingText={savingLabel(primary)}
-        disabled={isSaving && phase.savingAs !== primary}>
-        {saveLabel(primary)}
-      </Button>
-      <Button
-        onClick={() => onSaveAs(secondary)}
-        variant='outline'
-        className='w-full'
-        loading={phase.savingAs === secondary}
-        loadingText={savingLabel(secondary)}
-        disabled={isSaving && phase.savingAs !== secondary}>
-        {saveLabel(secondary)}
-      </Button>
-    </div>
-  )
-}
-
-/**
- * Default view when existing matches are found. Suppresses the parsed
- * card + save buttons so the user doesn't accidentally create a
- * duplicate. "Add anyway?" reveals the standard capture view.
- */
-function MatchesOnlyView({
-  matches,
-  onAddAnyway,
-}: {
-  matches: ExistingMatch[]
-  onAddAnyway: () => void
-}) {
-  return (
-    <div className='space-y-4'>
-      <SimilarMatchesList matches={matches} />
-      <Button onClick={onAddAnyway} variant='outline' className='w-full'>
-        Add anyway?
-      </Button>
-    </div>
-  )
-}
-
-function saveLabel(entityType: EntityType): string {
-  return entityType === 'contact' ? 'Save as contact' : 'Save as company'
-}
-
-function savingLabel(entityType: EntityType): string {
-  return entityType === 'contact' ? 'Saving contact…' : 'Saving company…'
-}
-
-/**
- * Folk's "N similar found" list. Each row pushes the corresponding detail
- * route, which fetches the full record via `record.getById`. The matched
- * field's value is shown as the row label — it's the most useful identifier
- * we have at this point in the flow (no displayName comes back from
- * `lookupByField`); the detail view fills in the rest.
- */
-function SimilarMatchesList({ matches }: { matches: ExistingMatch[] }) {
-  const { push } = useRouteStack()
-  const heading =
-    matches.length === 1 ? '1 similar record found' : `${matches.length} similar records found`
-  return (
-    <div className='space-y-2'>
-      <h3 className='text-xs font-medium text-muted-foreground'>{heading}</h3>
-      <ul className='space-y-1'>
-        {matches.map((match) => (
-          <li key={match.recordId}>
-            <button
-              type='button'
-              onClick={() =>
-                push({
-                  kind: match.entityType,
-                  recordId: match.recordId,
-                })
-              }
-              className='flex w-full items-center justify-between gap-2 rounded-md border border-border bg-muted/30 px-3 py-2 text-left text-sm transition-colors hover:bg-muted'>
-              <span className='min-w-0 flex-1'>
-                <span className='block truncate font-medium'>
-                  {match.entityType === 'contact' ? 'Contact' : 'Company'}
-                </span>
-                <span
-                  className='block truncate text-xs text-muted-foreground'
-                  title={match.matchedBy.value}>
-                  matched on {match.matchedBy.systemAttribute}: {match.matchedBy.value}
-                </span>
-              </span>
-              <ChevronRight className='size-4 shrink-0 text-muted-foreground' />
-            </button>
-          </li>
-        ))}
-      </ul>
-    </div>
-  )
-}
-
-function Empty({
-  title,
-  body,
-  action,
-}: {
-  title: string
-  body: string
-  action?: { label: string; href: string }
-}) {
-  return (
-    <div className='space-y-3'>
-      <h2 className='text-base font-medium'>{title}</h2>
-      <p className='text-sm text-muted-foreground'>{body}</p>
-      {action && (
-        <Button asChild className='w-full'>
-          <a href={action.href} target='_blank' rel='noreferrer'>
-            {action.label}
-          </a>
-        </Button>
-      )}
-    </div>
-  )
+  if (err instanceof Error) return err.message
+  return 'Save failed.'
 }
